Do not look for btree header byte
Looking for the header byte seemed like it worked but it didn't. The check
looking for it errored on '==' instead of '!=', so it turns out the header byte
was never really there. It was just a misreading of the sqlite file format,
then this bug in my code that made me think it was correct.